Refactor request into response.
